improve the performance of finding a palindrome with Task 13 v1.4

+++ 0x03-python-data_structures/13-is_palindrome.c
@@ -9,7 +9,7 @@ int is_palindrome(listint_t **head)
 {
 	listint_t *current = *head, *compare = *head;
 	unsigned int i, count_c = 0, count;
-	float sum2 = 0, sum = 0;
+	long double sum2 = 0, sum = 0;
 
 	for (i = 0; compare != NULL; i++)
 	{
